+++ backend/src/test/java/com/leetcode/monitor/controller/ContestIngestControllerTest.java
@@ -15,6 +15,7 @@ import com.leetcode.monitor.model.ScrapingStatus;
 import com.leetcode.monitor.parser.AcceptanceStatsParser;
 import com.leetcode.monitor.service.AcceptanceCalculationService;
 import com.leetcode.monitor.service.ContestStateService;
+import com.leetcode.monitor.service.ComparisonService;
 import com.leetcode.monitor.service.RankingService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -270,7 +271,8 @@ public class ContestIngestControllerTest {
         ContestStateService uninitializedService = new ContestStateService(
                 new AcceptanceStatsParser(),
                 new AcceptanceCalculationService(),
-                new RankingService()
+                new RankingService(),
+                new ComparisonService()
         );
         ContestIngestController controller = new ContestIngestController(uninitializedService);
         ContestIngestRequest request = new ContestIngestRequest("10 / 20", ScrapingStatus.SUCCESS);